feat(adapter): add factory for webhooks
