+++ src/adapters/supabase-factory.ts
@@ -17,7 +17,7 @@ import { SessionDataProvider } from './session/interfaces';
 import { SsoDataProvider } from './sso/interfaces';
 import { SubscriptionDataProvider } from './subscription/interfaces';
 import { ApiKeyDataProvider } from './api-keys/interfaces';
-import { IWebhookDataProvider } from './webhook/IWebhookDataProvider';
+import { IWebhookDataProvider } from '@/core/webhooks/IWebhookDataProvider';
 import { IOrganizationDataProvider } from '@/core/organization/IOrganizationDataProvider';
 
 
@@ -33,7 +33,7 @@ import { createSupabaseSessionProvider } from './session/factory';
 import createSupabaseSsoProvider from './sso/supabase/factory';
 import createSupabaseSubscriptionProvider from './subscription/factory';
 import createSupabaseApiKeyProvider from './api-keys/supabase/factory';
-import { createSupabaseWebhookProvider } from './webhook';
+import { createSupabaseWebhookProvider } from './webhooks';
 
 
 /**
