+++ src/adapters/webhooks/index.ts
@@ -1,27 +1,9 @@
+/**
+ * Webhook Adapter Exports
+ *
+ * Exposes webhook-related types and factory functions.
+ */
+
 export * from '@/core/webhooks/IWebhookDataProvider';
+export * from './factory';
 export * from './supabase/supabase-webhook.provider';
-
-import { IWebhookDataProvider } from '@/core/webhooks/IWebhookDataProvider';
-import { SupabaseWebhookProvider } from './supabase/supabase-webhook.provider';
-
-export function createSupabaseWebhookProvider(options: {
-  supabaseUrl: string;
-  supabaseKey: string;
-  [key: string]: any;
-}): IWebhookDataProvider {
-  return new SupabaseWebhookProvider(options.supabaseUrl, options.supabaseKey);
-}
-
-export function createWebhookProvider(config: {
-  type: 'supabase' | string;
-  options: Record<string, any>;
-}): IWebhookDataProvider {
-  switch (config.type) {
-    case 'supabase':
-      return createSupabaseWebhookProvider(config.options);
-    default:
-      throw new Error(`Unsupported webhook provider type: ${config.type}`);
-  }
-}
-
-export default createSupabaseWebhookProvider;
